+++ src/Enemies/UnitStat.cpp
@@ -4,11 +4,11 @@
 
 #include "UnitStat.h"
 
-UnitStat::UnitStat(const size_t statMax) {
+UnitStat::UnitStat(const unsigned int statMax) {
     setStatMax(statMax);
 }
 
-void UnitStat::setStatMax(const size_t statMax, const bool changeCurrent) {
+void UnitStat::setStatMax(const unsigned int statMax, const bool changeCurrent) {
     _statMax = statMax;
     if(changeCurrent) {
         _statValue = _statMax;
@@ -19,7 +19,7 @@ int UnitStat::getStatMax() const {
     return _statMax;
 }
 
-size_t UnitStat::decreaseStatBy(const size_t amount) {
+size_t UnitStat::decreaseStatBy(const unsigned int amount) {
     if(_statValue < amount) {
         size_t actualDamage = _statValue;
         _statValue = 0;
@@ -30,7 +30,7 @@ size_t UnitStat::decreaseStatBy(const size_t amount) {
     }
 }
 
-size_t UnitStat::increaseStatBy(const size_t amount) {
+size_t UnitStat::increaseStatBy(const unsigned int amount) {
     _statValue += amount;
     if(_statValue > _statMax) {
         size_t difference = _statValue - _statMax;
